Fix various integer size conversion warnings.
For buffer pre-allocations it is sufficient to silence the 
warning as those buffers would be expanded automatically, 
if they had to. But the system will run OOM before that.

* subversion/libsvn_fs_fs/fs_fs.c
  (rep_read_get_baton): silence conversion warning
  (fulltext_size_is_cachable): reject caching for huge items 
   on 32 bit systems
* subversion/libsvn_ra_svn/marshal.c
  (read_string): silence conversion warning
* subversion/libsvn_subr/io.c
  (svn_io_open_uniquely_named): dito
* subversion/libsvn_subr/cache-membuffer.c
  (NO_INDEX, NO_OFFSET): introduce proper unsigned defines
  (entry_t, svn_membuffer_t): adapt commentary
  (align_entry): replace
  (ALIGN_VALUE, ALIGN_POINTER): introduce alignment macros that 
   avoid conversion issues
  (drop_entry, insert_entry, get_group_index, find_entry, move_entry,
   membuffer_cache_set, membuffer_cache_get): use new defines
  (ensure_data_insertable, svn_cache__membuffer_cache_create): 
   use new defines; fix local variable types 

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/performance@992904 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/io.c
@@ -388,7 +388,7 @@ svn_io_open_uniquely_named(apr_file_t **file,
       /* Increase the chance that rand() will return something truely
          independent from what others get or do. */
       if (i == 2)
-        srand(apr_time_now());
+        srand((unsigned int)apr_time_now());
 
       /* Special case the first attempt -- if we can avoid having a
          generated numeric portion at all, that's best.  So first we
